feat: 'pathwar admin add-coupon' + 'pathwar admin coupons' + helpers to get id by slug and kind

+++ go/pkg/pwapi/api_agent-register.go
@@ -34,7 +34,7 @@ func (svc *service) AgentRegister(ctx context.Context, in *AgentRegister_Input) 
 	agent.Arch = in.Arch
 	agent.Version = in.Version
 	agent.Tags = strings.Join(in.Tags, ", ")
-	agent.NginxPort = in.NginxPort
+	agent.NginxPort = int64(in.NginxPort)
 	agent.Metadata = in.Metadata
 	agent.DomainSuffix = in.DomainSuffix
 	agent.AuthSalt = in.AuthSalt
